// types/form.ts
export enum ArtworkType {
  print = "Print",
  photo = "Photo",
  original = "Original",
  canvas = "Canvas",
  textile = "Textile",
  object = "Object",
}

export enum MountType {
  window = "Window",
  edge = "Edge",
  float = "Float",
  notSure = "Not Sure",
}

export enum GlassType {
  noGlass = "No Glass",
  standard = "Standard",
  antiReflective = "Anti Reflective / Anti UV",
  acrylic = "Acrylic",
  notSure = "Not Sure",
}

export enum WoodFinishType {
  paint = "Painted colour",
  black = "Blackened wood",
  white = "White dye",
  natural = "Natural Stain",
  waxed = "Waxed Wood",
  notSure = "Not Sure",
}

export enum WoodMaterialType {
  oak = "Oak",
  ash = "Ash",
  beech = "Beech",
  tulip = "Tulip",
  lime = "Lime",
  walnut = "Walnut",
  notSure = "Not Sure",
}